requirements.txt
cryptography==36.0.2
Flask==2.0.3
Flask_Cors==3.0.10
pycryptodome==3.14.1
PyJWT==2.3.0
pyodbc==4.0.32
